+++ tests/test_ntp_config.rb
@@ -1,7 +1,7 @@
 #
 # Minitest for NtpConfig class
 #
-# Copyright (c) 2015-2016 Cisco and/or its affiliates.
+# Copyright (c) 2015-2017 Cisco and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -26,28 +26,75 @@ class TestNtpConfig < CiscoTestCase
     no_ntpconfig
   end
 
+  def teardown
+    # teardown runs at the end of each test
+    no_ntpconfig
+    super
+  end
+
   def no_ntpconfig
     # Turn the feature off for a clean test.
     if platform == :ios_xr
-      config("no ntp source #{interfaces[0]}")
+      config("no ntp source #{interfaces[1]}")
     else
-      config("no ntp source-interface #{interfaces[0]}")
+      config("no ntp source-interface #{interfaces[1]}",
+             'no ntp authentication-key 111 md5 test 7',
+             'no ntp authentication-key 999 md5 test 7',
+             'no ntp trusted-key 111',
+             'no ntp trusted-key 999',
+             'no ntp authenticate')
     end
   end
 
   # TESTS
 
-  def test_create_destroy
+  def test_source_interface
     id = 'default'
 
     ntp = Cisco::NtpConfig.new(id)
     assert_includes(Cisco::NtpConfig.ntpconfigs, id)
     assert_equal(Cisco::NtpConfig.ntpconfigs[id], ntp)
 
+    assert_nil(Cisco::NtpConfig.ntpconfigs[id].source_interface)
+    assert_nil(ntp.source_interface)
+
     ntp.source_interface = interfaces[1]
     assert_equal(Cisco::NtpConfig.ntpconfigs[id].source_interface,
                  interfaces[1].downcase)
     assert_equal(Cisco::NtpConfig.ntpconfigs[id].source_interface,
                  ntp.source_interface)
+
+    ntp.source_interface = nil
+    assert_nil(Cisco::NtpConfig.ntpconfigs[id].source_interface)
+    assert_nil(ntp.source_interface)
+  end
+
+  def test_authenticate
+    id = 'default'
+
+    ntp = Cisco::NtpConfig.new(id)
+    assert_includes(Cisco::NtpConfig.ntpconfigs, id)
+    assert_equal(Cisco::NtpConfig.ntpconfigs[id], ntp)
+
+    assert_equal(false, Cisco::NtpConfig.ntpconfigs[id].authenticate)
+
+    ntp.authenticate = true
+    assert_equal(true, Cisco::NtpConfig.ntpconfigs[id].authenticate)
+  end
+
+  def test_trusted_key
+    id = 'default'
+
+    ntp = Cisco::NtpConfig.new(id)
+    assert_includes(Cisco::NtpConfig.ntpconfigs, id)
+    assert_equal(Cisco::NtpConfig.ntpconfigs[id], ntp)
+
+    assert_nil(Cisco::NtpConfig.ntpconfigs[id].trusted_key)
+
+    config('ntp authentication-key 111 md5 test 7',
+           'ntp authentication-key 999 md5 test 7')
+    ntp.trusted_key_set(true, 111)
+    ntp.trusted_key_set(true, 999)
+    assert_equal(%w(111 999), Cisco::NtpConfig.ntpconfigs[id].trusted_key)
   end
 end